(pages) move MainSearch into pages
